Packaging: don't delete debian dir in distclean
The kernel's built-in debian packaging deletes the ./debian directory in
distclean.  I don't see a fix for this upstream, but it was discussed in 2004:

	https://lkml.org/lkml/2004/8/14/133

Implement a portion of the suggested fix.  This patch doesn't fix the kernel's
built-in debian packaging, but it does prevent distclean from deleting the
debian directory.